Add files via upload
We can ensure that the code you write is functional and secure by testing it using whatever method you find best for your project. I used Junit testing here as I can test throughout my coding, and these are smaller projects. 
You interpret user needs and incorporate them by asking what your client's goal is. It would be best to communicate openly on needs, what the business wants and hopes will happen with the code, and then have users test it to ensure it exceeds expectations. You must think about the code outputs and inputs to function properly and everything that can go wrong and try those results.  I approach designing software by thinking of it as a blueprint. I start with my ideas on paper in a very rough form to get the basic hows and whys of the software. Then, after talking to the client and getting a better understanding, I put down the blueprint in ink and created a final design of what I wanted. Then, I start to work on the program but check in with the customer throughout the process to ensure it meets expectations. Then, could you finish testing the code and ensuring it runs properly and has the proper coverage?